Move the LMDB error handling into its own translation unit
lmdb_storage.cpp had grown a block of meta-page parsing and diagnostics formatting in the middle of the storage
implementation. Move raise_lmdb_exception, is_lmdb_corruption_error and LmdbEnvDiagnostics to
storage/lmdb/lmdb_error_handling.{hpp,cpp} so the storage impl reads as storage again. No behaviour change.

raise_lmdb_exception is now declared [[noreturn]], which it already was in effect. lmdb_extra_env_flags() stays in
lmdb_storage.cpp next to the mdb_env_open call it feeds, as it is not error handling.
