+++ cpp/arcticdb/storage/lmdb/lmdb_storage.hpp
@@ -12,8 +12,7 @@
 #include <arcticdb/util/pb_util.hpp>
 #include <arcticdb/storage/lmdb/lmdb_client_interface.hpp>
 #include <arcticdb/storage/lmdb/lmdb.hpp>
-#include <array>
-#include <cstdint>
+#include <arcticdb/storage/lmdb/lmdb_error_handling.hpp>
 #include <filesystem>
 #include <string>
 
@@ -31,38 +30,6 @@ struct LmdbInstance {
 /// changing library configs, e.g. to avoid an fsync per commit on CI where durability is irrelevant. Passed at open
 /// rather than mdb_env_set_flags because MDB_WRITEMAP can only be set at open.
 unsigned int lmdb_extra_env_flags();
-
-/// What LMDB believes about an env (mdb_env_info/stat) next to the two meta pages read straight from data.mdb with
-/// pread/ReadFile, bypassing the memory map. Attached to corruption-type errors so a failure in CI records whether
-/// the mapped view and the file disagree.
-struct LmdbEnvDiagnostics {
-    struct Meta {
-        uint32_t magic = 0;
-        uint32_t version = 0;
-        uint64_t mapsize = 0;
-        uint32_t psize = 0;
-        uint64_t free_root = 0;
-        uint64_t main_root = 0;
-        uint64_t last_pg = 0;
-        uint64_t txnid = 0;
-    };
-    unsigned int flags = 0;
-    size_t mapsize = 0;
-    size_t last_pgno = 0;
-    size_t last_txnid = 0;
-    unsigned int psize = 0;
-    unsigned int max_readers = 0;
-    unsigned int num_readers = 0;
-    std::array<Meta, 2> file_metas{};
-    std::string file_read_error;
-
-    std::string to_string() const;
-};
-
-/// True for LMDB error codes that indicate the env contents are not what LMDB expects
-bool is_lmdb_corruption_error(int error_code);
-
-LmdbEnvDiagnostics lmdb_env_diagnostics(::lmdb::env& env);
 
 class LmdbStorage final : public Storage {
   public:
